requirements.txt
defusedxml=0.4.1
Django=1.10.4
django-filter=1.0.1
django-revproxy=0.9.13
djangorestframework=3.5.3
djangorestframework-xml=1.3.0
gisdata=0.5.4
gsconfig=1.0.6
httplib2=0.9.2
psycopg2=2.6.2
requests=2.12.4
urllib3=1.16
